+++ lib/src/configuration/better_player_controls_configuration.dart
@@ -3,6 +3,7 @@ import 'dart:ui';
 
 // Flutter imports:
 import 'package:better_player/better_player.dart';
+
 // Project imports:
 import 'package:better_player/src/controls/better_player_overflow_menu_item.dart';
 import 'package:flutter/cupertino.dart';
@@ -107,11 +108,17 @@ class BetterPlayerControlsConfiguration {
   ///Flag used to show/hide qualities
   final bool enableQualities;
 
+  ///Flag used to show/hide PiP mode
+  final bool enablePip;
+
   ///Custom items of overflow menu
   final List<BetterPlayerOverflowMenuItem> overflowMenuCustomItems;
 
   ///Icon of the overflow menu
   final IconData overflowMenuIcon;
+
+  ///Icon of the PiP menu
+  final IconData pipMenuIcon;
 
   ///Icon of the playback speed menu item from overflow menu
   final IconData playbackSpeedIcon;
@@ -167,8 +174,10 @@ class BetterPlayerControlsConfiguration {
     this.enablePlaybackSpeed = true,
     this.enableSubtitles = true,
     this.enableQualities = true,
+    this.enablePip = true,
     this.overflowMenuCustomItems = const [],
     this.overflowMenuIcon = Icons.more_vert,
+    this.pipMenuIcon = Icons.picture_in_picture,
     this.playbackSpeedIcon = Icons.shutter_speed,
     this.qualitiesIcon = Icons.hd,
     this.subtitlesIcon = Icons.text_fields,