+++ lib/src/configuration/better_player_event_type.dart
@@ -17,4 +17,6 @@ enum BetterPlayerEventType {
   changedTrack,
   changedPlayerVisibility,
   changedResolution,
+  pipStart,
+  pipStop,
 }